Move `CGraph` class to separate compilation unit
