+++ CMakeLists.txt
@@ -2117,6 +2117,8 @@ if(CLIENT)
     friends.h
     ghost.cpp
     ghost.h
+    graph.cpp
+    graph.h
     graphics_defines.h
     graphics_threaded.cpp
     graphics_threaded.h
