+++ src/engine/client/client.h
@@ -23,6 +23,8 @@
 #include <engine/shared/network.h>
 #include <engine/warning.h>
 
+#include "graph.h"
+
 class CDemoEdit;
 class IDemoRecorder;
 class CMsgPacker;
@@ -39,32 +41,6 @@ class IUpdater;
 
 #define CONNECTLINK_DOUBLE_SLASH "ddnet://"
 #define CONNECTLINK_NO_SLASH "ddnet:"
-
-class CGraph
-{
-public:
-	enum
-	{
-		MAX_VALUES = 128,
-	};
-
-private:
-	float m_Min, m_Max;
-	float m_MinRange, m_MaxRange;
-	float m_aValues[MAX_VALUES];
-	float m_aColors[MAX_VALUES][3];
-	size_t m_Index;
-
-public:
-	void Init(float Min, float Max);
-	void SetMin(float Min);
-	void SetMax(float Max);
-
-	void Scale();
-	void Add(float v, float r, float g, float b);
-	void InsertAt(size_t Index, float v, float r, float g, float b);
-	void Render(IGraphics *pGraphics, ITextRender *pTextRender, float x, float y, float w, float h, const char *pDescription);
-};
 
 class CSmoothTime
 {
